move some methods into wizard component
